# openapi.yaml
openapi: 3.0.3
info:
  title: Word Learning App API
  description: API for a language word learning application with spaced repetition
  version: 1.1.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /topics/{topicId}:
    get:
      tags: [Topics]
      summary: Get topic with its subtopic IDs
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/topicId'
      responses:
        '200':
          description: Topic with subtopic IDs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── SUBTOPICS ──────────────────────────────────────────
...
  /admin/topics:
    get:
      tags: [Admin]
      summary: List all topics (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of topics with subtopic counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminTopic'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Admin]
      summary: Create a topic (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTopicRequest'
      responses:
        '201':
          description: Topic created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTopic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/topics/{topicId}:
    put:
      tags: [Admin]
      summary: Update a topic (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/topicId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminTopicRequest'
      responses:
        '200':
          description: Topic updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTopic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Admin]
      summary: Delete a topic (admin)
      description: Fails with 400 if the topic still has subtopics.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/topicId'
      responses:
        '204':
          description: Topic deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── ADMIN: SUBTOPICS ───────────────────────────────────
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Paginated list of users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: Page of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

# ─── COMPONENTS ─────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    topicId:
      name: topicId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller is authenticated but lacks required authority
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request payload or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Resource not found
...
    TopicDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        image_url:
          type: string
        subtopic_ids:
          type: array
          items:
            type: integer
            format: int64
          example: [1, 2, 3]
...
    Role:
      type: string
      enum: [USER, ADMIN]
      description: User authorization role. ADMIN unlocks /admin/** endpoints.
    AdminTopic:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        image_url:
          type: string
        sort_order:
          type: integer
        subtopics_count:
          type: integer
          format: int64
          example: 3
    AdminTopicRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
          example: Food
        description:
          type: string
          maxLength: 10000
        image_url:
          type: string
          maxLength: 255
        sort_order:
          type: integer
          minimum: 0
          example: 1
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        is_guest:
          type: boolean
        role:
          $ref: '#/components/schemas/Role'
        streak:
          type: integer
        gems:
          type: integer
        last_active_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
    AdminUsersPage:
      type: object
      properties:
        total:
          type: integer
          format: int64
          example: 137
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'